workin on it

+++ oss.h
@@ -3,8 +3,13 @@
 
 typedef struct Sh_mem {
 	int secs;
-	int nano_secs;
+	long nano_secs;
 	int prime_arr[20];	
+	//in prime_arr 0 means it didn't terminate
+	//positive is prime, negative is not 
+
+	
+
 
 } Sh_mem;
